Fix for crash in TeslaBeats
- PianoRoll should access it's corresponding timeline directly instead of asking for the global song's pattern timeline.

LMMS has a tighter coupling with the song where they have only one PianoRoll for the whole song and it's paired with the song's PlayPos for Pattern. Given the requirements for TeslaBeats, it ignores the song's PlayPos for Pattern and creates multiple piano rolls to facilitate live drum pattern recording. Ideally this should be fine; each PianoRoll creates it's own timeline widget and after that it sets the Song's pattern PlayPos to point to it. The song's pattern PlayPos will point to the last timeline that was set. This should be no problem as long as we don't access the pattern PlayPos from the song specifically. The piano roll, in one instance alone. Regardless of this tighter coupling of timeline + PlayPos in song, the right way for PianoRoll to access it's timeline is by referencing it's member directly.